web/backend/requirements.txt
asgiref==3.2.10
dj-database-url==0.5.0
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
Pillow==8.0.0
pytz==2020.1
sqlparse==0.4.1
whitenoise==5.2.0
